requirements.txt
streamlit==1.32.0
openai==1.12.0
diffusers==0.26.0
torch==2.2.0
torchvision==0.17.0
transformers==4.38.0
Pillow==10.2.0
opencv-python==4.9.0.80
moviepy==1.0.3
requests==2.31.0
python-dotenv==1.0.1
accelerate==0.27.0
numpy==1.24.3
scipy==1.11.4
tqdm==4.66.2
safetensors==0.4.2
